Organizando
Alterando estrutura de arquivos para deixar o código mais limpo.
